chore: Convert sheets on `/setting/appearance` route to use refs
